Create SampleWallet2.sol
contract Wallet { ... }: Defines a Solidity smart contract named "Wallet."

uint CurrentBalance;: Declares a state variable CurrentBalance of type uint to store the current balance of the smart contract.

function Deposit() public payable { ... }: Defines a function named Deposit that allows users to deposit funds into the wallet. The function is marked as payable to allow it to receive Ether (cryptocurrency).

CurrentBalance += msg.value;: Increases the CurrentBalance by the amount of Ether sent with the transaction.

function CheckBalance() public view returns (uint) { ... }: Defines a function named CheckBalance that allows users to check the current balance of the wallet. The function is marked as view since it does not modify the state.

return address(this).balance;: Returns the balance of the smart contract.

function SendTo(address payable to) public { ... }: Defines a function named SendTo that allows users to send the entire balance of the wallet to a specified address. The parameter to is marked as payable to indicate that it can receive Ether.

to.transfer(CheckBalance());: Transfers the entire balance of the smart contract to the specified address (to).